Added Object Poolng and cleaned up the scripts
As part of Object Pooling the instantiated bullets need to be reused and not deleted. From internal playtesting it was revealed that as the timing between bullets is 4 seconds, there will always be one bullet rendered on the map at all times. Therefore, only 1 bullet was being used and once it went off screen or collided with something it would stop travelling and stop being rendered. If the player presses Space the bullet will be moved to the player's position and angle and start the shooting mechanic.